broken. use FractSender.py for last running version.
Adding in fractal zoom of the datagrid.  finished step (1) - Grab the subgrid to be scaled, next; (2) - explode-scale into full size grid. then; (3) calculate heightmaps on the grid, algo will ignore existing values and fill in the gaps based on the existing values as seed.


same zooms require same rnd seed. currently fixed at 1. but not used, so is really sytem set as of now.